+++ Changes.md
@@ -5,6 +5,7 @@ refer to [https://github.com/msimerson/Mail-Toaster-6/commits/master](https://gi
 
 ## 2026-08
 
+- dma: create /var/spool/dma when the port has not
 - mt: jail control files move to /etc/mail-toaster/<jail>
 - mt: one /etc/mail-toaster/pfrule.sh serves every jail
 - pfrule: resolve rules under $MT6_ETC/<jail>/pf.conf.d